refactor: move geolocation button file as it will be used in two different places, it is now a generic component
