+++ src/features/search/components/SearchResults.tsx
@@ -9,7 +9,6 @@ import { ButtonContainer } from 'features/auth/signup/underageSignup/notificatio
 import { UseRouteType } from 'features/navigation/RootNavigator'
 import { Hit, NoSearchResult, NumberOfResults } from 'features/search/atoms'
 import { SingleFilterButton } from 'features/search/atoms/Buttons/FilterButton/SingleFilterButton'
-import { GeolocationButton } from 'features/search/atoms/Buttons/GeolocationButton'
 import { AutoScrollSwitch } from 'features/search/components/AutoScrollSwitch'
 import { ScrollToTopButton } from 'features/search/components/ScrollToTopButton'
 import { useHasPosition } from 'features/search/components/useHasPosition'
@@ -29,6 +28,7 @@ import { useIsFalseWithDelay } from 'libs/hooks/useIsFalseWithDelay'
 import { plural } from 'libs/plural'
 import { SearchHit } from 'libs/search'
 import { ButtonSecondary } from 'ui/components/buttons/ButtonSecondary'
+import { GeolocationButton } from 'ui/components/GeolocationButton'
 import { useHeaderTransition as useOpacityTransition } from 'ui/components/headers/animationHelpers'
 import { useModal } from 'ui/components/modals/useModal'
 import { HitPlaceholder, NumberOfResultsPlaceholder } from 'ui/components/placeholders/Placeholders'
