+++ src/ui/components/GeolocationButton/index.ts
@@ -1,0 +1,1 @@
+export { GeolocationButton } from './GeolocationButton'
